+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added `fileSize` to image and gallery fields
 - Added option to allow null values for `width`, `height` and `fileSize`
-- Added `maxLength` to text field
+- Added `characterLimit` to text field
 - Added text labels to `stylisedUi`
 - Renamed `ui` to `stylisedUi`
 - Renamed `size` to `fileSize`